+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Eflyt: Added functions for sending letters.
+
 ## [2.13.0] - 2025-08-13
 
 ### Added